Pass parameter to footer writer view open.
The footer will be written at close, possibly by the cache purge which
won't have as much knowledge about the file stream that it is closing
as the stack frame that opened the file stream. Thus, open now accepts
additional parameters that are passed to the footer writer when it is
invoked during close.

Closes #13.