Externalize `prepare` and `commit`.
This is primarily for unit testing, but why not give the user an
opportunity to verify an atomic file transaction themselves if they so
choose? Considered making this a separate object and may still consider
it, but instead decided to provide static function implementations of
the prepare and commit loops.

Closes #65.