Validity check directory and primary rename.
Validity check a directory rename as best you can, by simply seeing if
the directory is a directory. Validity check the rename of a file in the
primary directory by taking a hash when you generate the commit file for
the rename operation, as opposed to taking the hash during file write,
since you're not writing this file.

Closes #67.